create more weather module
